Separate delete and update endpoints
